+++ NEWS.md
@@ -1,3 +1,5 @@
+# incase (development version)
+
 # incase 0.3.2
 
 * Accept named arguments with or without preceding dots, e.g. `default` or `.default` (#5).