+++ NEWS.md
@@ -1,7 +1,8 @@
 # incase (development version)
 
-* `*_case()` functions can now use lists as return values.
-* `*_case()` functions no longer perform automatic type conversion on non-atomic return values.
+* `*_case()` functions can now use lists as return values (#7).
+* `*_case()` functions no longer perform automatic type conversion on non-atomic return values (#7).
+* `grep_case()` now supports vectors in the left-hand side of arguments (#8).
 
 # incase 0.3.2