+++ NEWS.md
@@ -1,4 +1,4 @@
-# incase (development version)
+# incase 0.4.0
 
 * The arguments `preserve`, `default`, and `ordered` have been deprecated in favor of `.preserve`, `.default`, and `.ordered` (#9).
   - Using undotted arguments will now trigger a warning. In future versions, these arguments will be removed.